Move occurrences to the nearest raster cell with data

Move point occurrences falling in raster cells without data (i.e. NA) to the nearest raster cell with data.

points2nearestcell(locs, ras, layer = 1, move = TRUE, distance = NULL,
  showchanges = TRUE, showmap = TRUE, leaflet = FALSE)

Arguments

locs
A SpatialPoints or SpatialPointsDataFrame.

ras
Raster* object.

layer
Integer. For RasterStack and RasterBrick objects, layer to use.

move
Logical. Change coordinates of points to those of the nearest raster cells?

distance
Numeric (optional). Maximum distance to move points. Point coordinates are only changed if the distance to the nearest raster cell is below distance.

showchanges
Logical. Print table with old and new coordinates.

showmap
Logical. Show map with original and new coordinates?

leaflet
Logical. If TRUE, show leaflet map instead of static map.

Value

A SpatialPointsDataFrame (with corrected coordinates if move is TRUE).

See also

https://github.com/SEEG-Oxford/seegSDM/blob/master/man/nearestLand.Rd and http://stackoverflow.com/questions/26652629/extracting-a-value-from-a-raster-for-a-specific-point-based-on-the-closest-cell/26688361#26688361.

Examples

# NOT RUN {
data(acaule)
locs <- locs2sp(acaule)
ras <- getData('worldclim', var = 'tmin', res = 10)
check.coords <- points2nearestcell(locs, ras)
# }